separated into Test and Main File
